fixing bugs and refactoring to make it work with create page
